Split archive file
